Added miasm-rs as a submodule
miasm-rs is a migration of many modules from Python to Rust, to improve speed and parallelization. At the moment, miasm-rs is not stable enough yet to become public. Only members that have access to the miasm-rs repository will be able to initialize this submodule.

When miasm-rs becomes public, this submodule will likely be replaced by a subtree, so no additional commands are needed to access the Rust version.

An alternative solution could have been to add miasm as a submodule to miasm-rs, however this would mean changing all links to Miasm files (examples, documentation...) in all external resources.